refactor: 네이버 oauth 정보 범위 축소
- 카카오와 맞추기 위해서 oauthId만 받아옴
