+++ src/main/java/com/somemore/global/auth/oauth/naver/dto/NaverUserProfileResponseDto.java
@@ -1,0 +1,18 @@
+package com.somemore.global.auth.oauth.naver.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public record NaverUserProfileResponseDto(
+        String resultcode, // 결과 코드
+        String message,    // 결과 메시지
+        Response response  // 응답 데이터
+) {
+    @JsonIgnoreProperties(ignoreUnknown = true)
+    public record Response(
+            String id // 일련 번호
+    ) {
+    }
+}
